Mini1-1706075016/testing/requirements.txt
gevent
greenlet
locustio
pyzmq
gevent-zeromq